# clients/shared/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Fatoura API",
    "description": "UAE VAT sales, purchases and invoicing.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/purchases": {
      "get": {
        "tags": [
          "Purchases"
        ],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "supplierId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfPurchaseSummaryDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Purchases"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PurchaseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurchaseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/purchases/{id}": {
      "get": {
        "tags": [
          "Purchases"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurchaseDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Purchases"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PurchaseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurchaseDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Purchases"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/purchases/{id}/attachment": {
      "get": {
        "tags": [
          "Purchases"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "Purchases"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "file"
                ],
                "type": "object",
                "properties": {
                  "file": {
                    "$ref": "#/components/schemas/IFormFile"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      },
      "delete": {
        "tags": [
          "Purchases"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
    "/api/reports/purchases": {
      "get": {
        "tags": [
          "Reports"
        ],
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "supplierId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurchasesReportDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DocumentCurrencyDto": {
        "required": [
          "code",
          "exchangeRate",
          "subTotalAed",
          "vatTotalAed",
          "totalAed"
        ],
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "exchangeRate": {
            "type": "number",
            "format": "double"
          },
          "subTotalAed": {
            "type": "number",
            "format": "double"
          },
          "vatTotalAed": {
            "type": "number",
            "format": "double"
          },
          "totalAed": {
            "type": "number",
            "format": "double"
          }
        },
        "description": "A document's currency, its rate (AED per unit) and its totals in AED, which reports and the VAT return use."
      },
...
      "IFormFile": {
        "type": "string",
        "format": "binary"
      },
...
      "PagedResultOfPurchaseSummaryDto": {
        "required": [
          "items",
          "total",
          "page",
          "pageSize"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PurchaseSummaryDto"
            }
          },
          "total": {
            "type": "integer",
            "format": "int32"
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "PartyDto": {
        "required": [
          "name",
          "address",
          "phone",
          "email",
          "trn"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "trn": {
            "type": "string"
          }
        }
      },
...
      "PurchaseDto": {
        "required": [
          "id",
          "number",
          "supplierInvoiceNo",
          "date",
          "supplierId",
          "supplier",
          "notes",
          "currency",
          "subTotal",
          "vatTotal",
          "total",
          "hasAttachment",
          "attachmentFileName",
          "createdAt",
          "lines"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "number": {
            "type": "string"
          },
          "supplierInvoiceNo": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "supplierId": {
            "type": "integer",
            "format": "int32"
          },
          "supplier": {
            "$ref": "#/components/schemas/PartyDto"
          },
          "notes": {
            "type": "string"
          },
          "currency": {
            "$ref": "#/components/schemas/DocumentCurrencyDto"
          },
          "subTotal": {
            "type": "number",
            "format": "double"
          },
          "vatTotal": {
            "type": "number",
            "format": "double"
          },
          "total": {
            "type": "number",
            "format": "double"
          },
          "hasAttachment": {
            "type": "boolean"
          },
          "attachmentFileName": {
            "type": [
              "null",
              "string"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "lines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PurchaseLineDto"
            }
          }
        }
      },
      "PurchaseLineDto": {
        "required": [
          "id",
          "lineNo",
          "itemId",
          "description",
          "expenseCategory",
          "quantity",
          "unitPrice",
          "taxCategory",
          "vatRate",
          "net",
          "vat",
          "total"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "lineNo": {
            "type": "integer",
            "format": "int32"
          },
          "itemId": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "description": {
            "type": "string"
          },
          "expenseCategory": {
            "type": "string"
          },
          "quantity": {
            "type": "number",
            "format": "double"
          },
          "unitPrice": {
            "type": "number",
            "format": "double"
          },
          "taxCategory": {
            "$ref": "#/components/schemas/TaxCategory"
          },
          "vatRate": {
            "type": "number",
            "format": "double"
          },
          "net": {
            "type": "number",
            "format": "double"
          },
          "vat": {
            "type": "number",
            "format": "double"
          },
          "total": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "PurchaseLineRequest": {
        "required": [
          "itemId",
          "description",
          "expenseCategory",
          "quantity",
          "unitPrice",
          "taxCategory"
        ],
        "type": "object",
        "properties": {
          "itemId": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "description": {
            "type": "string"
          },
          "expenseCategory": {
            "maxLength": 100,
            "type": [
              "null",
              "string"
            ]
          },
          "quantity": {
            "type": "number",
            "format": "double"
          },
          "unitPrice": {
            "type": "number",
            "format": "double"
          },
          "taxCategory": {
            "$ref": "#/components/schemas/TaxCategory"
          }
        },
        "description": "A purchase line: an item (adds stock at this unit cost) or an expense with a free-text category."
      },
      "PurchaseReportRow": {
        "required": [
          "id",
          "number",
          "supplierInvoiceNo",
          "date",
          "supplierName",
          "currency",
          "net",
          "vat",
          "total"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "number": {
            "type": "string"
          },
          "supplierInvoiceNo": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "supplierName": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "net": {
            "type": "number",
            "format": "double"
          },
          "vat": {
            "type": "number",
            "format": "double"
          },
          "total": {
            "type": "number",
            "format": "double"
          }
        },
        "description": "One purchase, in AED whatever its `Currency`."
      },
      "PurchaseRequest": {
        "required": [
          "supplierId",
          "supplierInvoiceNo",
          "date",
          "notes",
          "lines"
        ],
        "type": "object",
        "properties": {
          "supplierId": {
            "type": "integer",
            "format": "int32"
          },
          "supplierInvoiceNo": {
            "maxLength": 60,
            "type": [
              "null",
              "string"
            ]
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "notes": {
            "maxLength": 2000,
            "type": [
              "null",
              "string"
            ]
          },
          "lines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PurchaseLineRequest"
            }
          },
          "currency": {
            "maxLength": 3,
            "type": [
              "null",
              "string"
            ]
          },
          "exchangeRate": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          }
        }
      },
      "PurchasesBySupplier": {
        "required": [
          "supplierId",
          "supplierName",
          "count",
          "net",
          "vat",
          "total"
        ],
        "type": "object",
        "properties": {
          "supplierId": {
            "type": "integer",
            "format": "int32"
          },
          "supplierName": {
            "type": "string"
          },
          "count": {
            "type": "integer",
            "format": "int32"
          },
          "net": {
            "type": "number",
            "format": "double"
          },
          "vat": {
            "type": "number",
            "format": "double"
          },
          "total": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "PurchasesReportDto": {
        "required": [
          "period",
          "count",
          "net",
          "vat",
          "total",
          "bySupplier",
          "rows"
        ],
        "type": "object",
        "properties": {
          "period": {
            "$ref": "#/components/schemas/ReportPeriod"
          },
          "count": {
            "type": "integer",
            "format": "int32"
          },
          "net": {
            "type": "number",
            "format": "double"
          },
          "vat": {
            "type": "number",
            "format": "double"
          },
          "total": {
            "type": "number",
            "format": "double"
          },
          "bySupplier": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PurchasesBySupplier"
            }
          },
          "rows": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PurchaseReportRow"
            }
          }
        }
      },
      "PurchaseSummaryDto": {
        "required": [
          "id",
          "number",
          "supplierInvoiceNo",
          "date",
          "supplierId",
          "supplierName",
          "currency",
          "subTotal",
          "vatTotal",
          "total"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "number": {
            "type": "string"
          },
          "supplierInvoiceNo": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "supplierId": {
            "type": "integer",
            "format": "int32"
          },
          "supplierName": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "subTotal": {
            "type": "number",
            "format": "double"
          },
          "vatTotal": {
            "type": "number",
            "format": "double"
          },
          "total": {
            "type": "number",
            "format": "double"
          }
        }
      },
...
      "ReportPeriod": {
        "required": [
          "from",
          "to"
        ],
        "type": "object",
        "properties": {
          "from": {
            "type": "string",
            "format": "date"
          },
          "to": {
            "type": "string",
            "format": "date"
          }
        }
      },
...
      "TaxCategory": {
        "enum": [
          "Standard",
          "ZeroRated",
          "Exempt"
        ]
      },
...
    }
  },
...
}